refactor(auth): consolidate email provider + make Provider.service optional
A Provider can now be auth-only (no media-service counterpart) or
service-only (no auth) or both — the service field is optional. Email
and Discord move from fake-service-shell to clean auth-only providers.
Service queries filter on p.service presence via a type guard; the ad-hoc
AUTH_ONLY_PROVIDER_IDS list is gone.

Email login logic moves from packages/backend/src/auth/providers/email.ts
into packages/backend/src/providers/email/index.ts so all providers share
a single folder layout. The auth/ directory is removed (was holding one
file). /api/auth/login + /api/auth/register keep importing registerEmail
/ loginEmail from the new location.
